Rename test files accordingly
